start lifting TDL
