+++ packages/E2ETest/windows/ReactUWPTestApp/ReactPackageProvider.cs
